Cleanup on amber validation.
- change KeyError to interface.
- Combine ExamplesInteractiveServer test files into one.
- Breakout mandatory pattern test into its own file.
- Fix grammar with optional key missing MismatchMessages.
