+++ core/src/main/kotlin/io/specmatic/core/KeyError.kt
@@ -2,15 +2,17 @@ package io.specmatic.core
 
 import io.specmatic.core.Result.Failure
 
-sealed class KeyError {
-    abstract val name: String
+sealed interface KeyError {
+    val name: String
 
-    abstract fun missingKeyToResult(keyLabel: String, mismatchMessages: MismatchMessages = DefaultMismatchMessages): Failure
+    fun missingKeyToResult(keyLabel: String, mismatchMessages: MismatchMessages = DefaultMismatchMessages): Failure
 
-    abstract fun missingOptionalKeyToResult(keyLabel: String, mismatchMessages: MismatchMessages = DefaultMismatchMessages): Failure
+    fun missingOptionalKeyToResult(keyLabel: String, mismatchMessages: MismatchMessages = DefaultMismatchMessages): Failure {
+        return missingKeyToResult(keyLabel, mismatchMessages).copy(isPartial = true)
+    }
 }
 
-data class MissingKeyError(override val name: String) : KeyError() {
+data class MissingKeyError(override val name: String) : KeyError {
     override fun missingKeyToResult(keyLabel: String, mismatchMessages: MismatchMessages): Failure =
         Failure(mismatchMessages.expectedKeyWasMissing(keyLabel, name))
 
@@ -19,7 +21,7 @@ data class MissingKeyError(override val name: String) : KeyError() {
     }
 }
 
-data class UnexpectedKeyError(override val name: String) : KeyError() {
+data class UnexpectedKeyError(override val name: String) : KeyError {
     override fun missingKeyToResult(keyLabel: String, mismatchMessages: MismatchMessages): Failure =
         Failure(mismatchMessages.unexpectedKey(keyLabel, name))
