+++ core/src/test/kotlin/io/specmatic/core/pattern/JSONObjectPatternTest.kt
@@ -1083,9 +1083,9 @@ internal class JSONObjectPatternTest {
             assertThat(result).isInstanceOf(Result.Failure::class.java)
             assertThat(result.reportString()).containsIgnoringWhitespaces("""
             >> topLevelOptionalKey
-            Expected Optional key named "topLevelOptionalKey" was missing
+            Expected optional key named "topLevelOptionalKey" was missing
             >> subMandatoryObject.subOptionalKey
-            Expected Optional key named "subOptionalKey" was missing
+            Expected optional key named "subOptionalKey" was missing
             >> subOptionalObject.subOptionalKey
             Expected key named "subOptionalKey" was missing
             """.trimIndent())
