+++ core/src/test/kotlin/io/specmatic/core/pattern/ListPatternTest.kt
@@ -202,7 +202,7 @@ Feature: Recursive test
     }
 
     @Test
-    fun `should not result in failure when list is empty but pattern is cycling and failure when not and resolver is set to allPatternsAsMandatory`() {
+    fun `should not result in failure when list is empty but pattern is cycling and resolver is set to allPatternsAsMandatory`() {
         val basePattern = ListPattern(parsedPattern("""{
             "topLevelMandatoryKey": "(number)",
             "topLevelOptionalKey?": "(string)",
@@ -211,7 +211,7 @@ Feature: Recursive test
         """.trimIndent(), typeAlias = "(baseJsonPattern)"), typeAlias = "(baseListPattern)")
         val listPattern = ListPattern(basePattern)
 
-        val matchingValue = parsedValue("""[
+        val value = parsedValue("""[
             [
                 {
                     "topLevelMandatoryKey": 10,
@@ -226,19 +226,11 @@ Feature: Recursive test
             ]
         ]
         """.trimIndent()) as JSONArrayValue
-        val matchingResult = listPattern.matches(matchingValue, Resolver(newPatterns = mapOf("(baseListPattern)" to basePattern)).withAllPatternsAsMandatory())
-        println(matchingResult.reportString())
-        assertThat(matchingResult).isInstanceOf(Result.Success::class.java)
-        assertThat(matchingResult.reportString()).isEmpty()
+        val result = listPattern.matches(value, Resolver(newPatterns = mapOf("(baseListPattern)" to basePattern)).withAllPatternsAsMandatory())
 
-        val nonMatchingValue = JSONArrayValue(matchingValue.list.plus(JSONArrayValue(emptyList())))
-        val nonMatchingResult = listPattern.matches(nonMatchingValue, Resolver(newPatterns = mapOf("(baseListPattern)" to basePattern)).withAllPatternsAsMandatory())
-        println(nonMatchingResult.reportString())
-        assertThat(nonMatchingResult).isInstanceOf(Result.Failure::class.java)
-        assertThat(nonMatchingResult.reportString()).containsIgnoringWhitespaces("""
-        >> [1]
-        List cannot be empty
-        """.trimIndent())
+        println(result.reportString())
+        assertThat(result).isInstanceOf(Result.Success::class.java)
+        assertThat(result.reportString()).isEmpty()
     }
 
     @Test
