+++ core/src/test/resources/openapi/tracker.yaml
